BUG 9744: Reworking batch requests in Rails Multiple threads (I believe) could result in overlapping calls to batch_request in different blocks. These would get aggregated into a single call, but then the return value wouldn't be in the proper context, resulting in lost data and crashes.
So rework it to make each batch request block take an id,
which is then passed in to the core server requests that should
be batched in that block. This ensures the proper set of batched
requests are processed together.

Then fix up all the places that use it, which also involves
adding a parameter to update & delete requests for that batch id
(get and create already had a boolean that was appropriated
for the id).

Also rearranged things a bit when actually calling the batch
to have safer handling of the queue; however, this is not as
important with the above rework.

QA: Site should generally load/work same as normal. Specific cases
to check:
- The Grouped Table canvas.test page
- Profile page, with the numbers in the upper-right
- Profile, deleting Open ID links from an account
- Profile, checking/unchecking items to email you about
- Routing & Approval section of About Dataset
- Internal panel, checking/unchecking features
- Internal panel, deleting properties
- Administration, creating & deleting categories
- Admin, viewing and managing routing & approval

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>