FEATURE 8815: Properly display sub-categories for datasets * Refactor View.categories into tree * In edit metadata, render under parent with -- * In browse & govstat Manage Data, render tree only if parent or child   is selected; just render top-level category otherwise
QA: Create a category with a parent of another category in the Metadata
section of Admin. When editing metadata for a dataset, you should be
able to select and save that category. In the dropdown, it should appear
under its parent category, and have '--' before it to indicate a
sub-level.
When browsing data, by default only top-level categories are shown in
the sidebar. If you select the parent category, it should show the
children of that category. Selecting that child category should still
show the expanded tree. A dataset assigned to a child category will
appear in browse for both the child and parent category.
Similar behavior should work on GovStat Manage Data.

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>