BUG 9896: Embedded maps don't zoom via scroll by default New configuration for map which disables scroll-zoom by default. It listens for any mouseup in the map, and enables scroll-zoom. Use this property for maps in DataSlate.
QA: Maps embedded in DataSlate pages, such as the goal page on govstat
and canvas.test, get new behavior. By default, when you use the
mousewheel to scroll through the page, the map will no longer catch it
and zoom. Instead, you must first click anywhere in the map; and then
scrolling to zoom will be enabled for that map. Maps in widgets and
dataset page should behave same as before (scroll to zoom is on by
default).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>