FEATURE 10013, BUG 9893: user-less frontend html output When a user is logged in we often embed the username and user object in the page. For caching pages which can be shared between users we need to change this so javascript gets the current user and performs the appropriate page changes client-side.
This is the first stage of this transition; removing the userId and
User objects.

Execute a call to the core server to get the current user as soon
as possible. Setup a function to which callbacks can be stored as
soon as a response is available. Once received, set blist.currentUser*

QA: Full regression across most features within the site. This could
   impact virtually anything which uses the current logged in user or
   user id. Pay attention to nominations, widgets, and discussions
   in particular; and keep a keen eye on the upper right "sign in",
   "Hello, [Your Name]" and look at the js console closely for any
   errors. Nasty responses from IE are not completely unexpected.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>