BUG 10029: Option to enable/disable 'All' for Hawaii layout * 'show_all' option in gov_stat config. Defaults to true * Make new gov_stat config, and prefer it for dashboard_layout, too * Move a useful Class function into Configuration.rb
QA: Changing the layout of a GovStat page should still work normally. By
default, the list layout on GovStat should show the 'All' tab, as
before. Once you have toggled the govstat layout once (to make sure a
new configuration is created), you can go to the Internal section for
the domain, choose the GovStat option under 'Other Configurations', and
add a value with key of 'show_all' and value of 'false'. This should
cause the 'All' tab on the list layout to be hidden.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>