FEATURE 9761: Header component * Inherits from Picture * Centers the image, maxes out the width, allows a height to be set that clips, adds a border below * If present, hides default header in print * Set defined height in Rails * Handle height as String
QA: In govstat report builder, you should be able
to add a Header component. This takes an image,
centers it on the page, and adds a border beneath it.
You can specify a height, which will clip the image
to that height; but the picture will be squished to fit
if it is wider than the page. If you have a header
on the page, print preview should show the default
govstat header hidden.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>