BUG 10020: When updating a DataSet, account for no active rows being present.
The high-level cause of this bug was us noticing an old-style layer
references in the related views, which are populated when various sidebar
buttons are pressed (such as "Discuss"). We then try to do a _convertLegacy
but die because we haven't populated the active row set.
Fix is to skip processing the active row set if we don't have such a set
yet.

QA: Ensure bug is fixed. Additionally, ensure migration of old Maps
views still works properly (views that have no uid='self').

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>