FEATURE 10090, 10087: Localization for categories * Add 'available_locales' key to locales configuration so we know what   languages we should support on the domain. Defaults to just an array   of the current locale. * Categories now store 'locale_strings' for each item. This stores an   optional translation for each locale. If no value is present for a   locale, it falls back to the primary value. * Admin: Metadata now has fields for each locale (if more than one) and   handles using them. In particular, an update option was added to   modify existing values. * When getting categories for category_tree, it automatically selects   just the text for the current locale. * New method 'category_display' was added on View to display the   appropriate string.
QA: Categories, including updating them in admin, should behave as
before on single-language domains.
To make a domain multi-lingual: Go to the internal panel for it, and add
a configuration of type 'locales'. In that config, first add a property
for which languages should be available:
  available_locales: ["en", "fr"]
Then add a property for the current language for the domain:
  *: fr
(the key is just the single character '*').
Now if you go into Admin: Metadata, under the categories section you
should get a new text box for each locale. You should be able to update
or create categories with different strings for the locales. The correct
string for the current domain locale should be rendered wherever that
category is shown on screen: browse (both sidebar and in table view),
About this Dataset, Edit Metadata, etc.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>