BUG 10151: Properly handle no hidden categories Only happened if there less than facet cutoff number of categories (5), and a non-existant category was chosen as a param.
QA: Browse should work as normal. Test a domain that has less than 5
total categories. Click on one of the categories to filter by it. Then
change the URL so it says 'category=nonexistant' to filter by a category
that is not in the list. The page should load (and show no results).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>