BUG 10691: Don't return a bogus value in _updateDataSource.
In short, we were returning true when we had no new contexts to process.
This confused us later on, so we ended up trying to update the map
instead of setting it up for the first time.

QA: Ensure bug is fixed. Also general pass over DataSlate pages with
maps/charts.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>