BUG 8789: Fix grid-sidebar not letting its neighbor know the new size.
QA: This was an insidious issue inherent in all grid-sidebar pages (basically,
pages with the familiar sidebar full of configuration options). I've changed
how grid-sidebar interacts with the resize event, so please check that resizing
the window works with such pages, as well as showing and hiding the sidebar.
Please make sure to check for script error messages, as I've added some sanity
checking to the code that should print out errors if things don't go the way
I expected.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>