BUG 10867: D3 Bar: Resolve timing issue (size comes in before data).
Issue was that we incorrectly assumed that we'd get a resize when we
get data, but that's not necessarily true. So we'd skip the resizing
logic too zealously. Fix is to do the resizing logic as much as is
possible with partial info. Simple change, though.

QA: Ensure bug is fixed.

Reviewed-by: Michael Chui <michael.chui@socrata.com>