BUG: Correctly handle array params in browse.js.
QA: Go to here on any site:

/browse/select_dataset?limitTo[]=charts&limitTo[]=maps&limitTo[]=blob&sortBy=alpha&suppressed_facets[]=type&utf8=✓

It should only show charts, maps, and blobs (that behavior isn't part
of this fix). Now, change the filters, and especially try sorting and
searching. Ensure that it continues to show charts, maps, and blobs.

Also make sure browse works as usual in general.

To-be-
Reviewed-by: jeff.scherpelz@socrata.com