BUG 10911: D3 Bar: Don't go to negative scale for small viewport.
Clamp the scale to positive, and hide the legend under certain extreme
sizing pressures.

QA: Ensure bug is fixed. Also make sure legends in general continue
to behave the same as before, except that they hide to give more
room if the chart is <140px in the y-axis direction. We should only
hide when we reserve space for the legend, namely these configurations:
Column chart: Legend top or bottom, and height < 140px.
Note that our minimum width prevents this from being an issue with
bar charts; we'll just cut off content if they get too thin. If it weren't
for the minimum width, we'd hide the legend under these situations as well:
Bar chart: Legend left or right, and width < 140px (impossible today).

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>