FEATURE 10586: Charts: Allow config of DSG colors in UI.
QA: If you go and enable DSG in the sidebar, you'll see
a new Colors config option. It will color the virtual
columns - please verify that it works, and that the order
of the colors in the configuration matches the order of colors
in the legend. Note that this is active for old charts as well,
so please do a quick check of old charts with this option.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>