FEATURE 11213: Header changes
QA: The 'i' button in headers no longer has a flyout that says 'Column Info' when hovered on. Instead, it shows the flyout with all column information on hover. The message should disappear on any click outside of the flyout, when the mouse leaves the flyout or when any other flyout is shown.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>