BUG 11416, 11094: D3 Charts DSG: Don't miss rows, batching of processing, progress UI.
We don't lock up the browser as bad now! Translations ony for en as of this
commit.

QA: Bugs should be fixed. In addition, you'll now see a progress screen when
new charts are calculating series grouping, complete with a pause/resume
button. Please play with this button a bit (perhaps changing the chart
configuration while paused or running, etc). Also a regression pass over
DSG in general.

Reviewed-by: Michael Chui <michael.chui@socrata.com>