BUG 11434: Accessible star-rating widget * Rendered as form, inputs, and labels in HTML * Mostly CSS to make it look like stars * Keyboard accessible with & without JS * Add text-focus style * Submit button and Rails controller for non-JS * JS is much simplified, but can still update a control with the current value if needed * Added a class to skip uniform() in basePane
QA: Star-rating widget should still look and behave as normal. If you do
not have Javascript enabled, you should get a Save button that saves it
via Rails. You should also be able to tab to it, and use the left-right
arrows on the keyboard to change the value. In IE8, the editable stars
are just radio buttons due to compatibility.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>