BUG 11751: D3 Line/Bar: Alphabetize Groups in sidebar
QA: This commit adds a checkmark (Alphabetize Groups) to the sidebar under
Data Series Grouping.
This needs to be thoroughly tested on a variety of series grouped charts.
It should cause the order of the bars and the order of legend entries to
be alphabetized according to the value of the column you're grouping by.
More specifically:
If you have just one data column, the legend/bars should be in regular
alphabetical order. So take a bar chart with this checkmark active.
Your data column is, say, "Expenditures", and you're grouping by another
column, "Department". So in this case, the legend will show department names
in alphabetical order, regardless of any sorts applied to the original
dataset and regardless of the order of the data entered into the dataset.
So your legend could show, to make up some department names:
1) Accounting
2) Billing
3) Engineering
4) Management
Note the alphabetical order. The legend order should match the column order,
as usual.

Now let's add another data column, "Account Balance". Now the legend (and
bar order) should look like this:
1) Expenditures, Accounting
2) Account Balance, Accounting
3) Expenditures, Billing
4) Account Balance, Billing
5) Expenditures, Engineering
6) Account Balance, Engineering
7) Expenditures, Management
8) Account Balance, Management

Note that the legend is in alphabetical order on the series column value,
NOT the order of the data columns. So basically it's alphabetical on the
first letter after the comma. This general idea should extend to all
charts with the checkmark enabled.

Disabling the checkmark should go back to the old behavior with no change.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>