BUG 11704: D3 Label in bar: Handle nulls/invalid data correctly.
QA: Bug should be fixed. Null or invalid values should cause no labels to
be displayed within the bars (OK to display the regular way outside the
bars, if so configured). For stacked, only the label (not the value) should
be drawn inside the stacked bars, but only if at least one part of the
stack isn't null. Completely null stacks should show no labels inside of
them.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>