BUG 11863: Making a change to a DSG chart triggers the new default teal instead of maintaining the original dark Blue
QA: Look at a previously created chart with dsg. The dsg colors will not match the chart colors. If you uncheck dsg, click apply and then check it and click apply again, the chart's colors should correspond to the dsg colors, however  the first color should still be still dark blue. Main point is that the first of the new color palette is still the same as the old dark blue.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>