FEATURE 11879: Fix up Report Builder interactions * Split Edit into Select, and then Edit * Select allows resize, remove, move (drag/drop) * Edit hides those, focuses on component, and allows text editing, sidebar config * Add icon next to remove to enter edit mode * Click on selected, or double-click on unselected goes to edit * Esc goes from edit to select to unselected; Enter goes from selected to edit * Fixed up base control to handle multiple controls per dom * Show move cursor in select mode to indicate drag/drop is available * Prevent container from being dropped inside itself * Fix up bug where a container would hang on to drop position past when the item left
QA: Interactions in Report Builder should be clearer. You can still drag
components around when none are selected; but now when you select a
component, you will see the resize handle and remove button as before;
but you can still drag it around in this mode (but not edit text or
properties). Instead, click again, or click the new 'Edit' button next
to 'Remove', or hit Enter to go into edit mode. This will grey out the
other components, hide the Remove & Edit buttons and resize handle, but
enable text editing and properties. You cannot drag the component around
in this mode. You can also hit Esc to go from edit mode back to the
selected mode, and Esc one more to completely deselect the component.

There are also some small fixes with dragging around containers
(horizontal & vertical layouts): you shouldn't be able to drag and drop
a container into itself; and if you drag a component out of a container
and drop it, then drag it back into the container and drop it, it should
work.

Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>