BUG 11578: D3 Charts: Variable legend width with legend on left/right
QA: Full pass on legends in all d3 charts (but only a light pass on pie
and donut). Make sure to test with X and Y axis labels present and not
present. When placed on the left or right, the legend's width should
reflect the width of its content. So long legend entries should stretch
the legend up to 35% of the chart width, and short entries should allow
it to collapse. In addition, legends with many entries should now be
scrollable in left or right positions. Please verify this as well.

Reviewed-by: Michael Chui <michael.chui@socrata.com>