BUG 11991: New Charts: bar chart axis lines overlap legend
QA: The legend should now be floating above the tick lines, and indeed
everything else in the chart. A quick pass over charts in general to make
sure the legend still looks right in the main scenarios. Note that the
legend is semitransparent, so you'll still see the tick marks through
the legend. This is expected.

Reviewed-by: Michael Chui <michael.chui@socrata.com>