BUG: PERF: Table view - don't accidentally relayout the entire window on hover
This is just some low-hanging fruit. There are systemic issues still in table.js,
chief of which is the fact that we completely replace DOM elements via innerHTML
instead of updating the elements already there. This makes FF and IE slowwww
(chrome is ok).

QA: Make sure row highlighting and focusing/editing still works in tables.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>