BUG CORE-53: D3 Charts, Column charts: Labels in bars appear cut off
QA: Bug should be fixed. I wasn't able to repro the axis baseline jumping
off screen anymore, but I did fix some labels disappearing in IE8 and I now
allow the topmost label in a stacked bar or column to keep going past the
bar, but only if the text is black.

Reviewed-by: Michael Chui <michael.chui@socrata.com>