BUG CORE-62: (NYS) Charts: Don't orphan flyouts on re-render.
QA: This issue is triggered when there's a chart rerender while there's a
flyout on-screen. This is tricky to do but one way is to choose a large
dataset, then mouse over columns as we're pulling in data. The issue
manifests itself as stuck flyouts that don't disappear even after pressing
"Apply" again. Please also do a quick pass around charts in DataSlate pages,
especially pages that change the chart in response to a button press or other
user interaction.

Reviewed-by: Michael Chui <michael.chui@socrata.com>