BUG CORE-62: (NYS) Charts: AutoUpdateSort.
QA: This introduces a checkmark to preserve sort order when you click
apply in a pie chart sidebar. If this is disabled, it will keep whatever
sort was previously set (we used to always sort descending on the first
value column, paving over whatever was there). Note that this doesn't
resolve the Other slice issue also mentioned in the bug; that's going to
be a separate commit on Tuesday.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>