BUG: New Visualize - Misc fixes. Min angle for pie + axis decimals are now back to sliders. If conditional formatting is active a warning is shown in the color section. Fewer checkboxes are shown in Legend config.
QA: Check so that the sliders for decimals and min angle works. Check so that if conditional formatting is active a message shows up in the color section, make sure these links work (clear formatting and link to Conditional formatting tab). Lastly, new behavior for checkboxes in legend config. If you have value markers these should be displayed by default in the legend. This is also true for conditional formatting, for this one  if there is no conditional formatting, then the conditional formatting checkbox shouldn't even show up in legend config.

Reviewed-by: Gus Glover <gus.glover@socrata.com>
Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>