FEATURE STAT-142: Enable rich list and page view in DataSlate Visualize * Move Rails pieces into templates, and load them via RTM * Fix up ids -> classes (so we don't duplicate in DataSlate) * Remove some ie7 & ie6 styles * Fix up Rails issue with loading Page displayType * Load dataset-controls with rich renderers, table/dataset styles, too * Load translations in RTM
QA: In Report Builder you should be able to use the Visualize component
to add a view that has been saved with a rich list or page view, and it
should render properly. Also general check on views/charts/maps with
toggling on and off the display of map/chart/table/rich list/page.

Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>