FEATURE: Handle saved views in SODA2 * A view is always based off of a parent dataset or view. When the   parent has a query, that query _must_ be applied first. If a user does   not have rights to edit a view, then any filtering is applied on top   of the saved view. So: * Dataset now has queryBase to tell it view to apply against. This view   is used for all row requests. This may be the current view/dataset   (but the saved/unmodified version). * Subtract out anything from the queryBase, and just apply the   difference from the current view. * Split blist.viewCache to also have blist.sharedViewCache.   blist.viewCache should truly be an un-modified cached copy of the   view; the modified one goes in sharedViewCache. * Filter sidebar now shows static version of queryBase, and just allows   modification on top of that. * Keep filter metadata through translation and back; also don't simplify   (remove single-child join operators and empty items) for filter sidebar * Check for metadata.hideBase flag on root condition to hide base filter section
QA: Filtering on views should work as normal, along with editing the
default filter. The filter sidebar will now show what the base query is,
which you can't modify; you will only be able to add queries on top of
that. If you are the owner of a view, you should be able to modify the
query. But if you're an unsigned user, then the saved query on the view
should be un-editable, and you can only make changes on top of it.

Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>