BUG STAT-314: Show federation in dataset picker.
QA: In goal edit page, federated datasets should show a "Provided-by" link
on the right hand side of the dataset listing. In addition, the dataset
popup that shows up on hovering over a dataset should show a similar
link. If the dataset is not federated, these should not appear. Please
check the behavior across a few resolutions - the text should not overlap.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com