BUG CORE-1018: Fix up a couple filtering bugs * Trim out blank conditions from baseRootCondition. This occurs when a   condition just has metadata, so it should show a prompt to the user.   That should stay in the primary root condition, not in the base. * Fix bug where a single filter condition would get simplified to a   point the filter sidebar couldn't handle it. If that is the case,   re-add the necessary parents.
QA: Create a view and set up a default filter in the sidebar with
nothing selected (just options for a user to choose). Look at the view
as an unsigned user, and make sure you still see the prompts to filter.
Set up a filter sidebar as above with only one colum and operator, and
no values selected. Save that view. Then as the owner, add a filter
value and save it. Make sure the condition doesn't disappear.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>