BUG CORE-1032: Fix case of bad data in dataset metadata under certain conditions Background: a dataset (view) stores which render types (table, page, map, etc.) are visible; but it also has a separate key for which dataset ids are active for those views. In particular, this handles the case where a map has multiple layers based on different datasets, and the user wants to see one of the other datasets in the grid.
Problem: If none of the ids under the active set are the current
dataset, then rows will never be loaded and the current dataset will
never be marked as ready, in particular for the sidebar to show. This
happened when there was a map with a non-current dataset table, and then
the map was hidden. This resulted in the sidebar never showing.

Solution: In _adjustProperties, detect this state in the metadata, and
force one of the visible items to be the current view.

QA: Create a map with a second (different) datasets as another layer.
Then choose to show the table/grid for that other dataset. Then unselect
the map, leaving just the table visible. Save that view, and reload:
make sure the sidebar still opens.

Reviewed-by: Michael Chui <michael.chui@socrata.com>