FEATURE CORE-1179: HTTP basic endpoint for domain cache invalidation * Flags the current domain as out-of-date * Skips cookie auth, uses basic auth instead * Also requires the user to have edit_site_theme, edit_pages, or create_pages right
QA: There is a new endpoint /admin/flag_out_of_date (you'll have to go
there directly, there is no link in the UI for it). It won't work even
if you're already logged in; your browser will prompt you for username
and password. It should deny access unless you are an admin or publisher
(maybe an editor, too?; definitely not normal users, though). If it
works, it will just say "Done" in the browser.

Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>