FEATURE CORE-1325: Verify that we can quickly turn off SODA2 on a domain on which it is enabled.
The intention of this change is to allow one of two URL parameters to be added to a URL to force the front-end to use SODA v1 or SODA v2 API when making XHR calls.

QA: Please test the following use cases:

1) Using a domain that does not have the "use_soda2" feature flag set, view a dataset or view, add the URL parameter "soda=2" to the URL and verify that SODA v2 API calls are being used.
2) Using a domain that does have the "use_soda2" feature flag set, view a data set and alter the URL to include "&soda=1" to switch to using SODA v1 XHR calls.

You will be able to verify that the XHR requests use the old API since they will no longer include the "version=2.0" URL parameter.

Example V1 URL looks something like this:
https://.../views/INLINE/rows.json?accessType=WEBSITE&method=getByIds&asHashes=true&start=0&length=50&meta=true

Example V2 URL looks something like this:
https://.../api/id/vdj8-z7h2.json?%24%24exclude_system_fields=false&%24offset=0&%24limit=34&%24%24version=2.0&_=1390349959611

Reviewed-by: Michael Chui <michael.chui@socrata.com>