BUG CORE-1345: Upgrade Hashie gem from v0.3.1 to v2.0.5
A bug was recently discovered in the Hashie gem that we're using in the Rails front-end. The version we're using is over 3 years old and an update is long overdue.

QA: While automated tests and manual testing in my development environment did not reveal any issues. You may consider doing a full regression if time permits.

Reviewed-by: Michael Chui <michael.chui@socrata.com>