BUG STAT-562: Better handle private views in narratives.
QA: (Same as janus-govstat message)
Bug should be fixed. Given I'm editing a narrative, when I add a private
dataset to the narrative, I should see the dataset but also see a piece
of red text below it telling me it's private. As a non-signed-in user,
viewing a narrative with a private dataset should show a cross-hatched
area where the dataset should be, and also a "Dataset is private" text
at the bottom.

Reviewed-by: Jason Kroll <jason.kroll@socrata.com>