BUG CORE-1561: Cannot add row to dataset with soda2
The request when sent to the server using the SODA2 API was failing due to a malformed request.

After fixing the request format, the response from the server uses a different data structure that wasn't being handled in the JS front-end.

This change remedies both issues and has been tested against OBE and NBE.

QA: This can be tested by visiting the following URL in staging and adding multiple new rows to the dataset: https://budget-automated-tests.test-socrata.com/dataset/Capital-Project-Sample-Data-For-Mo-Co/dg9j-sguf

Reviewed-by: Michael Chui <michael.chui@socrata.com>