TASK CORE-1622: Make frontend proxy $$store to core
This changes the frontend (both Rails and Javascript code) to pass through the "$$store=" parameter to the backend when making requests for dataset rows.

Note: This is not an exhaustive modification. There may be other areas of code requiring changes in order to pass through this parameter for other kinds of requests.

QA: Test by accessing the following URL:

https://soda2obe.test-socrata.com/Government/Chicago-Crimes-10k/hpp5-cajj?$$store=pg.primus

Verify that browser loads data into the grid and there are no JavaScipt errors.

Reviewed-by: John Kew <john.kew@socrata.com>