NOOP: Downtime for Azure + Production
Azure is having a special set of maintenance windows this upcoming
weekend, with staging happening in advance of it.

Production is scheduled for 3/15, and per the plan, we're likely to
shorten it from 4 hours to 2 (making each maintenance window this
quarter be a 2 hours, for the same total downtime.)

QA: You'll see the new downtime message starting later today through
Wednesday.

Reviewed-by: Michael Chui <michael.chui@socrata.com>