FEATURE CORE-1708: UX Deprecation for SODA Server datasets in Q1
This changes the frontend to disable the edit button for NBE datasets until this functionality is provided in future revisions.

Included among these changes:

  o  Additional unit testing
  o  Minor refactoring of View model
  o  Addition of logDebug() method to debug.js
  o  Addition of "Read Only Data" messaing to en.yml
  o  Adjustments to Gemfile to allow debugging within RubyMine
  o  Addition of URL parameter to allow override of read-only datasets ($$force_editable=true)
  o  Refactoring of inline logic in dataset/show view to use helper methods
  o  Alteration of dataset.js to add preliminary support for adding rows to NBE datasets (enclose JSON payload in [])

QA: Test by verfiying that OBE datasets can be edited by creating a working copy. Verify that NBE datasets do not display the Edit button. Verify that that adding the $$force_editable=true to the URL while viewing an NBE dataset causes the Edit button to appear. Note that NBE datasets do not yet support making working copies, so that is expected to fail.

Reviewed-by: Michael Chui <michael.chui@socrata.com>

Includes revisions from code review.