# This is a combination of 2 commits. # The first commit's message is:
FEATURE CORE-1614 : Frontend must communicate if data from secondary is out-of-date.

Closely related to CORE-1613.

Up-to-dateness is presented in the grid. Displays the text, "This dataset is out of date by N seconds" in the event that the dataset is behind truth by N seconds.

QA: This cannot be tested until other changes are deployed.

Reviewed-by: John Kew <john.kew@socrata.com>
Reviewed-by: Michael Chui <michael.chui@socrata.com>