FEATURE CORE-1705 : UX changes to prevent SOQL merging in Q1
This change completely disables the Filter pane when the dataset is 1) new backend and 2) of view type "tabular" (via the is_blist? method).
Also factored out nearly all of the if/this/that/or/the/other/thing logic in the show view into helper methods so we can both read and test the code.
Added "non_tabular?" method to view class to represent the all-too-common usage of "!is_tabular?" and changed said usages to use "non_tabular?
Factored out much of the repetitive code in the show view that was used to build links, into helper methods for readability and testability.
Short-circuited a conditional in show view that always evaluated to true (newChartsEnabled).

QA: Please carefuly regression test the "pill" links across the top of the grid view to ensure that they all function normally and are visible / invisible in the appropriate cases and that the panes that they cause to appear in the sidebar appear and function properly. Also, please use a dataset that contains a Formatted Text column and confirm that when editing the value of one of these cells, that the formatting buttons that appear above the grid function as they are expected too.

Reviewed-by: Michael Chui <michael.chui@socrata.com>