NOOP CORE-1705 : FIX UX changes to prevent SOQL merging in Q1
This reverses the logic for when to show/hide the Filter pane (and sub-panes of the Manage pane) when the dataset is new backend and "tabular".

QA: The Filter pane should appear only WHEN the dataset is new backend and is tabular (non-view, etc.). The original commit for this work got the logic backwards.

Reviewed-by: Michael Chui <michael.chui@socrata.com>