BUG CORE-1809: Frontend Changes for Case Sensitivity
This change alters the work done in CORE-1809 to conditionally wrap the string matching where clause filters only if the dataset is new backend (because "soql.lobotomize = true" in Core config in production).

QA: Confirm that filtering of text columns ignores upper/lower case of string values provided to "IS", "IS NOT" and "STARTS WITH" filter functions in both Old backend dataset and New backend datasets. When testing Old backend datasets, please also modify the URL to include "&soda=1" or "&soda=2" and verify that the filters still behave normally.

Reviewed-by: John Kew <john.kew@socrata.com>