BUG CORE-1883: Dataset out-of-date notification does not appear when it should
This change fixes a bug that caused the out-of-date message to not appear when SODA2 headers specified that the dataset was out of sync with truth.
The message is now internationalized.
The message is now propagated via a new dataset event (dataset_last_modified).
The message is now defaulted to "The dataset is up-to-date as of ..."
The times are formatted using the MomentJS library.
Default MomentJS date format added (blist.configuration.shortDateFormat and blist.configuraiton.longDateFormat).
Displays messaging for plain dataset as well as working copies (which are how we identify unpublished datasets).

QA: Verify new backend datasets display either "This dataset is up-to-date as of <last modified>" or "This dataset is out of date by <age of dataset>" message below the dataset name in the top section of the grid view. Note that synthesizing out-of-date datsets may be difficult (or impossible) in staging, so the first case may be the only one that is reasonble to verify.

Reviewed-by: Michael Chui <michael.chui@socrata.com>