FEATURE CORE-1971 : Export from DataSpace is broken if you filter first.
This change disables the Download feature for new backend datasets that have been filtered.
A message is displayed instead directing the user to download the dataset from which the view is derived.

QA: Please verify that Export -> Download is only available for "plain" new backend datasets (i.e. not filtered views, etc.). Availability of Download should be unaffected for old backend datasets.

Reviewed-by: Michael Chui <michael.chui@socrata.com>