FEATURE CORE-1818: Add feature flag editing UI to dataset edit_metadata
QA: If you are a superadmin, you should see a new "Feature Flags" in /edit_metadata.
If you are not, you should not see this.
You should be able to turn on and off the feature flags on a per-dataset basis from this.
This can be verified at /4x4/flags.data (Testing the feature flags is strictly optional.)

Reviewed-by: Randy Antler <randy.antler@socrata.com>