BUG: CORE-2007 More Views side panel blank for Baltimore open data set upon interaction
Defense against map usage and empty metadata.  A view that hits this break more views/comments tabs in the whole group.

QA: Overlay two maps of the same dataset.  Use the API to set metadata of the map to null.  Go back to the default dataset.  Clicking more views should not break.

Reviewed-by: michael.chui@socrata.com