BUG CORE-1535: Make it possible to delete both Schrödinger's Metadata fields.
Repro steps:
1. Create a new field in /admin/metadata. Let's call it foo/bar.
2. Set it to private.
3. Go to dataset D. Edit metadata. Set foo/bar to "private".
4. Go back to /admin/metadata. Set foo/bar to public.
5. Go back to dataset D. Edit metadata. Set foo/bar to "public".

Expected: For both signed and unsigned users, foo/bar is "public".
Actual: For unsigned users, foo/bar is "public". For signed users, foo/bar is "private".

Root cause:
1. Our idea of hiding private metadata is putting it in a different hash in the view data. In the database, this is lenses.private_metadata and lenses.metadata.
1a. The reasoning for this is probably that there's no easy way for the backend to cull private metadata on the fly depending on user permissions. Regardless, the backend arguably does the correct thing here. It could be better.
2. The frontend assumes that, if it receives privateMetadata, then the privateMetadata must be correct and should replace the public metadata.
3. Therefore, it simply does a blind merge.
4. The actual result is that dataset D in the repro example shows D.metadata.custom_fields.foo.bar as "public" and D.privateMetadata.custom_fields.foo.bar as "private".
5. In theory, the latter property would be deleted when executing repro step 4, but that's extremely unreasonable.

Implemented fix:
Check the contents of /admin/metadata for the privacy setting and force View#merged_metadata to respect it.

Note:
This doesn't attempt to fix the quantum state of a given metadata field. Thus, if a metadata field has had public/private set to it, swapping the privacy setting will actually select the other value. This behavior is still problematic, but it is *better*.

QA: Perform the repro steps, and then delete the fields by saving something blank to them.

Reviewed-by: Randy Antler <randy.antler@socrata.com>