BUG: Column chart sometimes displays bars with wrong height when filtering.
QA: Go to any new UX page. Filter on a bar. The resultant clicked bar
should always get smaller or stay the same, but never get bigger.
This is hard to repro, so try on a variety of browsers. I was only able to
get it to repro once, but it's a severe bug (displays wrong data), so it's
worth the extra effort.
Reviewed-by: randy.antler@socrata.com