BUG: Column chart expand button showed up inappropriately.
QA: Go to any page which has a column chart with few columns. Make the
window big so that the columns reach their maximum allowed size (30px).
The show more button should not show up at the bottom-right of the chart.
Also, ensure that for a column chart with many bars (like chicago crime's
crime type), that the show more button shows up if the window is small
enough.
Reviewed-by: randy.antler@socrata.com